utils/intermediate_reps/requirements.txt
lark==1.1.9
graphviz==0.20.3
PyYAML==6.0.1
